Implement the Automaton trait for FSTs.
Adds an implementation of Automaton on FSTs, which makes it possible
to pass an FST into Map.search or Set.search to, for example, search
a map for a large set of keys all at once, rather than one at a time,
taking advantage of the existing efficient state machine construction
implementation for sets. The result seems to be faster for large sets
than looking up one at a time, and is more memory-efficient than the
current regex implementation, and *much* faster than the current
intersection operation.